+++ include/SegmentFitting_reg.hxx
@@ -28,21 +28,21 @@ protected:
     // inpute
     const CDataTempl<UINT8>  *m_pSem_bg;  // [2D: ht*wd]
     const CDataTempl<UINT32> *m_pSemI;    // [2D: ht*wd]
-    const CDataTempl<float>  *m_pSemMat;  // [3D: ht*wd*num_classes]
-    const CDataTempl<float>  *m_pDistMat; // [3D: ht*wd*num_dist]
+    const CDataTempl<double>  *m_pSemMat;  // [3D: ht*wd*num_classes]
+    const CDataTempl<double>  *m_pDistMat; // [3D: ht*wd*num_dist]
     UINT8 m_fitseg_method;
     
     // generated data, used in whole class.
     UINT32 m_ht,  m_wd;
     UINT32 m_num_sem, m_num_dist;
-    CDataTempl<float>     m_lutX;
+    CDataTempl<double>     m_lutX;
 
     // parameter
     const GlbParam     *m_pParam;
     
     
 public:
-    Segment_Fit(const GlbParam *pParam, const CDataTempl<UINT8> *pSemBg, const CDataTempl<UINT32> *pSemI, const CDataTempl<float> *pSemMat, const CDataTempl<float> *pDistMat, UINT8 fit_method=1){
+    Segment_Fit(const GlbParam *pParam, const CDataTempl<UINT8> *pSemBg, const CDataTempl<UINT32> *pSemI, const CDataTempl<double> *pSemMat, const CDataTempl<double> *pDistMat, UINT8 fit_method=1){
         m_pParam   = pParam;
         m_pSem_bg  = pSemBg;
         m_pSemI    = pSemI;
@@ -71,13 +71,13 @@ protected:
     // find key points given point list.
     void FindKeyPoints(const UINT32 numPt, const vector<UINT32> &ptY, const vector<UINT32> &ptX, UINT32 *dist_ch, vector<UINT32> &out_key_idxs);
     // attribute of segment starting at key_idxs[stK]:
-    vector<vector<float> > FitErrorStartFrom(UINT32 dist_ch, float meanY, UINT32 len, UINT32 stK,
+    vector<vector<double> > FitErrorStartFrom(UINT32 dist_ch, double meanY, UINT32 len, UINT32 stK,
                                     const vector<UINT32> &ptY, const vector<UINT32> &ptX, const vector<UINT32> &key_idxs);
     void SemanticScoreStartFrom(auto &semScore, UINT32 len, UINT32 stK, 
                                     const vector<UINT32> &ptY, const vector<UINT32> &ptX, const vector<UINT32> &key_idxs);
     
     // Dynamic Programming to find best partitions of the line.
-    void DP_segments(const CDataTempl<float> &seg_info, auto &semScore, const vector<UINT32> &ptY, const vector<UINT32> &ptX,
+    void DP_segments(const CDataTempl<double> &seg_info, auto &semScore, const vector<UINT32> &ptY, const vector<UINT32> &ptX,
                       const vector<UINT32> &key_idxs, vector<UINT32> &dp_idxs);
 };
 
@@ -85,7 +85,7 @@ void Segment_Fit::InitLUTx(UINT32 len){
     m_lutX.Init(len, 6); // 0-acc_x^2, 1-acc_x, 2-1/(acc_x^2*(x+1) - acc_x*acc_x)
                            // 3-1/(x+1), 4-1/(x^2), 5-x
     m_lutX.SetData(1.0, 0, 3);
-    float cumsum_powK = 0, cumsum_K = 0;
+    double cumsum_powK = 0, cumsum_K = 0;
     for(UINT32 k = 1; k < len; k ++){
         cumsum_K    += k;
         cumsum_powK += pow(k, 2); 
@@ -102,7 +102,7 @@ void Segment_Fit::InitLUTx(UINT32 len){
 void Segment_Fit::SemanticScoreStartFrom(auto &semScore, UINT32 len, UINT32 stK, 
                     const vector<UINT32> &ptY, const vector<UINT32> &ptX, const vector<UINT32> &key_idxs){
     UINT32 idx_len = key_idxs.size();
-    vector<float>  acc_sem_score(m_num_sem, 0);
+    vector<double>  acc_sem_score(m_num_sem, 0);
     for(UINT32 k =stK+1; k < idx_len; k ++){
         Mkey_2D key(stK, k);
         semScore[key].resize(m_num_sem, 0);
@@ -111,7 +111,7 @@ void Segment_Fit::SemanticScoreStartFrom(auto &semScore, UINT32 len, UINT32 stK,
             for(UINT32 i=key_idxs[k-1]; i < key_idxs[k]; i++){
                 acc_sem_score[j] += m_pSemMat->GetData(ptY[i], ptX[i], j);
             }
-            semScore[key][j] = acc_sem_score[j] / float(pt_len);
+            semScore[key][j] = acc_sem_score[j] / double(pt_len);
         }
     }
 }
@@ -134,14 +134,14 @@ void Segment_Fit::SemanticScoreStartFrom(auto &semScore, UINT32 len, UINT32 stK,
  *     ptY/ptX: all points.
  *     key_idxs: index of key points.
  */
-vector<vector<float> > Segment_Fit::FitErrorStartFrom(UINT32 dist_ch, float meanY, UINT32 len, UINT32 stK,
+vector<vector<double> > Segment_Fit::FitErrorStartFrom(UINT32 dist_ch, double meanY, UINT32 len, UINT32 stK,
                                             const vector<UINT32> &ptY, const vector<UINT32> &ptX, const vector<UINT32> &key_idxs){
     // compute acc_y, acc_xy, acc_y_2
-    vector<float> acc_y(len, 0);
-    vector<float> acc_xy(len, 0);
-    vector<float> acc_y_2(len, 0);
+    vector<double> acc_y(len, 0);
+    vector<double> acc_xy(len, 0);
+    vector<double> acc_y_2(len, 0);
 
-    float distV = m_pDistMat->GetData(ptY[key_idxs[stK]], ptX[key_idxs[stK]], dist_ch);
+    double distV = m_pDistMat->GetData(ptY[key_idxs[stK]], ptX[key_idxs[stK]], dist_ch);
     acc_y[0]    = distV;
     acc_xy[0]   = 0;
     acc_y_2[0]  = pow(distV, 2);
@@ -154,10 +154,10 @@ vector<vector<float> > Segment_Fit::FitErrorStartFrom(UINT32 dist_ch, float mean
 
     // compute w and b for segments starting at stK and ending at latter key points
     UINT32 idx_len = key_idxs.size();
-    vector<vector<float> > fit_err(idx_len, vector<float>(3, 0.0));
+    vector<vector<double> > fit_err(idx_len, vector<double>(3, 0.0));
     for(UINT32 k =stK+1; k < idx_len; k ++){
         UINT32 tk = key_idxs[k] - key_idxs[stK]; // length of segments [point_stK, point_k]
-        float w, b;
+        double w, b;
         if (m_fitseg_method == 1){
             b = (acc_y[tk]*m_lutX.GetData(tk, 0) - acc_xy[tk]*m_lutX.GetData(tk, 1)) * m_lutX.GetData(tk, 2);
             w = (acc_xy[tk] - b*m_lutX.GetData(tk, 1)) * m_lutX.GetData(tk, 4);
@@ -167,11 +167,11 @@ vector<vector<float> > Segment_Fit::FitErrorStartFrom(UINT32 dist_ch, float mean
             b = (acc_y[tk] - w*m_lutX.GetData(tk, 1)) * m_lutX.GetData(tk, 3);
         }
 
-        float term_0  = w*w*m_lutX.GetData(tk, 0) + 2*w*b*m_lutX.GetData(tk, 1) + b*b*(m_lutX.GetData(tk, 5)+1);
-        float term_1  = 2*(w*acc_xy[tk] + b*acc_y[tk]);
-        float err_sum = term_0 - term_1 + acc_y_2[tk];
+        double term_0  = w*w*m_lutX.GetData(tk, 0) + 2*w*b*m_lutX.GetData(tk, 1) + b*b*(m_lutX.GetData(tk, 5)+1);
+        double term_1  = 2*(w*acc_xy[tk] + b*acc_y[tk]);
+        double err_sum = term_0 - term_1 + acc_y_2[tk];
         err_sum       = err_sum>=0? err_sum : -err_sum;
-        fit_err[k][0] = err_sum/max(min(acc_y[tk], meanY * (m_lutX.GetData(tk, 5)+1)), float(1e-9));
+        fit_err[k][0] = err_sum/max(min(acc_y[tk], meanY * (m_lutX.GetData(tk, 5)+1)), double(1e-9));
         fit_err[k][1] = w;
         fit_err[k][2] = b;
     }
@@ -183,7 +183,7 @@ vector<vector<float> > Segment_Fit::FitErrorStartFrom(UINT32 dist_ch, float mean
  * compute fitting error for combination of any two key points.[st, end] 
  */
 void Segment_Fit::FittingFeasibleSolution(Fit_Mode mode, Segment_Stock *pSegStock){
-    auto ComputeMeanDistance = [&](UINT32 *dist_ch, UINT32 numPt, vector<UINT32> &ptY, vector<UINT32> &ptX, float *meanD){
+    auto ComputeMeanDistance = [&](UINT32 *dist_ch, UINT32 numPt, vector<UINT32> &ptY, vector<UINT32> &ptX, double *meanD){
         meanD[0] = 0;   meanD[1] = 0;
         for(UINT32 k=0; k < numPt; k++){
             meanD[0] += m_pDistMat->GetData(ptY[k], ptX[k], dist_ch[0]);
@@ -226,20 +226,20 @@ void Segment_Fit::FittingFeasibleSolution(Fit_Mode mode, Segment_Stock *pSegStoc
             ptXs.assign(m_ht+1, k);
 
         // prepare.
-        float meanD[2];
+        double meanD[2];
         ComputeMeanDistance(dist_ch, num_pt, ptYs, ptXs, meanD);
         vector<UINT32> key_idxs;
         FindKeyPoints(num_pt, ptYs, ptXs, dist_ch, key_idxs);
        
         // compute information for each small segment.
         UINT32 num_key = key_idxs.size();
-        CDataTempl<float> segInfo(num_key, num_key, 5);
-        map<Mkey_2D, vector<float>, MKey2DCmp>      semScore;
-        map<Mkey_2D, map<string, float>, MKey2DCmp> fitResult;
+        CDataTempl<double> segInfo(num_key, num_key, 5);
+        map<Mkey_2D, vector<double>, MKey2DCmp>      semScore;
+        map<Mkey_2D, map<string, double>, MKey2DCmp> fitResult;
         for(UINT32 j=0; j < num_key-1; j++){
             UINT32 len_2end = num_pt - key_idxs[j];
-            vector<vector<float> > fit_err_0 = FitErrorStartFrom(dist_ch[0], meanD[0], len_2end, j, ptYs, ptXs, key_idxs); 
-            vector<vector<float> > fit_err_1 = FitErrorStartFrom(dist_ch[1], meanD[1], len_2end, j, ptYs, ptXs, key_idxs); 
+            vector<vector<double> > fit_err_0 = FitErrorStartFrom(dist_ch[0], meanD[0], len_2end, j, ptYs, ptXs, key_idxs); 
+            vector<vector<double> > fit_err_1 = FitErrorStartFrom(dist_ch[1], meanD[1], len_2end, j, ptYs, ptXs, key_idxs); 
             SemanticScoreStartFrom(semScore, len_2end, j, ptYs, ptXs, key_idxs);
             for(UINT32 i=j+1; i<key_idxs.size(); i++){
                 segInfo.SetData(fit_err_0[i][0]+fit_err_1[i][0], j, i, 0);
@@ -311,7 +311,7 @@ void Segment_Fit::FindKeyPoints(const UINT32 numPt, const vector<UINT32> &ptY, c
 /*
  * Dynamic Programming to find best partitions of the line.
  */
-void Segment_Fit::DP_segments(const CDataTempl<float> &seg_info, auto &semScore, const vector<UINT32> &ptY, const vector<UINT32> &ptX,
+void Segment_Fit::DP_segments(const CDataTempl<double> &seg_info, auto &semScore, const vector<UINT32> &ptY, const vector<UINT32> &ptX,
                       const vector<UINT32> &key_idxs, vector<UINT32> &dp_idxs){
     auto NewSegmentFromSemantic = [&](UINT32 k, UINT32 st_k, UINT32 len){
         // semantic background
@@ -324,7 +324,7 @@ void Segment_Fit::DP_segments(const CDataTempl<float> &seg_info, auto &semScore,
         if(k < len-1){
             Mkey_2D obsKey(k, k+1);
             Mkey_2D expKey(st_k, k);
-            float diff = _ChiDifference(semScore[obsKey], semScore[expKey]);
+            double diff = _ChiDifference(semScore[obsKey], semScore[expKey]);
             if(diff > m_pParam->segFit_dp_semdiff_thr)
                 return 2;
         }
@@ -334,8 +334,8 @@ void Segment_Fit::DP_segments(const CDataTempl<float> &seg_info, auto &semScore,
 
     // forward-backward recordings:
     UINT32 len  = key_idxs.size();
-    vector<float>  acc_costV(len, 0);
-    vector<float>  seg_costV(len, 0);
+    vector<double>  acc_costV(len, 0);
+    vector<double>  seg_costV(len, 0);
     vector<int>    acc_numV(len, 0);
     vector<int>    bk_routeV(len, 0);
 
@@ -355,14 +355,14 @@ void Segment_Fit::DP_segments(const CDataTempl<float> &seg_info, auto &semScore,
         
         // traverse all availabe selection, find the optimal solution
         UINT32 min_idx = 0;
-        float min_cost = 2*m_pParam->segFit_dp_inf_err, min_fitCost=0, min_numSeg=0;
+        double min_cost = 2*m_pParam->segFit_dp_inf_err, min_fitCost=0, min_numSeg=0;
         for(UINT32 j = st_k; j < k; j ++){
             // compute cost for having segment [iniIdxs[st_k],  iniIdxs[j])
-            // float BIC_cost  = (acc_numV[j]+1) * Log_LUT(m_iniIdxs[k]-m_iniIdxs[st_k]+1);
-            float BIC_cost  = (acc_numV[j]+1) * log(key_idxs[k]-key_idxs[st_k]+1);
-            float fit_cost  = seg_costV[j] + seg_info.GetData(j, k, 0)*(key_idxs[k]-key_idxs[j]);
-            float fit_resist = m_pParam->segFit_dp_inf_err * (seg_info.GetData(j, k,0) > m_pParam->segFit_dp_err_thr);
-            float cand_cost  = fit_cost + m_pParam->segFit_dp_bic_alpha*BIC_cost + fit_resist;
+            // double BIC_cost  = (acc_numV[j]+1) * Log_LUT(m_iniIdxs[k]-m_iniIdxs[st_k]+1);
+            double BIC_cost  = (acc_numV[j]+1) * log(key_idxs[k]-key_idxs[st_k]+1);
+            double fit_cost  = seg_costV[j] + seg_info.GetData(j, k, 0)*(key_idxs[k]-key_idxs[j]);
+            double fit_resist = m_pParam->segFit_dp_inf_err * (seg_info.GetData(j, k,0) > m_pParam->segFit_dp_err_thr);
+            double cand_cost  = fit_cost + m_pParam->segFit_dp_bic_alpha*BIC_cost + fit_resist;
 
             if(cand_cost < min_cost){
                 min_cost    = cand_cost;
